#41 Shooting AI should be done now. Added a small debugging utitlity to visually see what the AI is targeting. Maybe we will figure out something prettier later, but it works for now. Updated Matrix4 to be column-major. Attempted replacing Matrix4x4 with Matrix4, but there was some odd results when combined with view frustum culling.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@167 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Utility/Quaternion.cpp
@@ -227,15 +227,15 @@ void Quaternion::getMatrix(GLfloat* matri)
 
 Matrix4 Quaternion::getMatrix()
 {
-	double x2 = x * x;
-	double y2 = y * y;
-	double z2 = z * z;
-	double xy = x * y;
-	double xz = x * z;
-	double yz = y * z;
-	double wx = w * x;
-	double wy = w * y;
-	double wz = w * z;
+	float x2 = x * x;
+	float y2 = y * y;
+	float z2 = z * z;
+	float xy = x * y;
+	float xz = x * z;
+	float yz = y * z;
+	float wx = w * x;
+	float wy = w * y;
+	float wz = w * z;
     
    Matrix4 matrix(
   			1.0f - 2.0f * (y2 + z2), 2.0f * (xy - wz), 2.0f * (xz + wy), 0.0f,
